+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * c278bfd - 2018-07-26: Pre-Release v9.3.56.dev4 
+- * 774481c - 2018-07-27: Fixed file browser modal 
+- * 24b4751 - 2018-07-26: Pre-Release v9.3.56.dev4 
 - * 334b7c7 - 2018-07-26: Misc JS code fixes 
 - * 96a1978 - 2018-07-26: Fixed IMDb star ratings for display show view 
 - * 9c41f4b - 2018-07-26: Fixed genre row alignment for display show view Fixed small poster sizing for shows view